Switch to MinGW compiler
I have finally decide to switch to a more flexible compiler, I have also updated the CMakeLists.txt file, and fixed a typo in the .gitignore. Some code refactors were needed to cope with the new compiler, but all is setup for now. I have also added some new brief documentation for some of the changes.