Switched to an ECS structure
I have finally switched to an Entity-Component-System structure, which lets us abstract a lot of the code away. It was hard refactoring all of this, but I think the code cleanliness and possible optimizations make it totally worthwhile. It still needs to be polished, but it's going greatly. Also, whilst testing, I have discovered Issue #2, though I also hope to find a fix for it ASAP.