Updated GLM to iuse Git
I noticed GLM, GLFW, TinyOBJLoader, Tracy, and the stb_image.h libraries weren't actually using Git, so they appeared as part of the main project, and they weren't updating. I'm fixing this with a few commits, strating with this one. I'm making them all independent for two reasons, cleanliness, and the fact that GitKraken struggles if the commits are too big.