Added support for per-model texturing
It's not the best, but it's not bad! There are probably some things left to do, and I would like to revamp the ECS to support component dependencies, as to not go wrong, i.e., adding a texture component to an entity that doesnt have a model component, etc. Anyways, it has shaders, its own render system, it supports per-model textures, and plays fairly well with the rest of the code. For now, the "floor" quad is the only thing that defaults to having a texture. The system can also be extended to have per-model uniform buffers, though I still have to work that out.